Change: add functionality to api.
Add client.bucketKeys to return a promise that is fufilled with all
keys.

Add client.bucketDeleteAll to remove all keys in a bucket. This returns
a promise which is fufilled once all the keys have been removed. Note
that is very slow since it waits for all deleted keys to be reaped. See
http://lists.basho.com/pipermail/riak-users_lists.basho.com/2012-May/008433.html
for details.